pyproject.toml
[tool.poetry]
name = "hn_han"
version = "2.14.0"
description = "Basic reading of information from Aidon HAN ports."
authors = ["John Åsmund Westberg <thejaws@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
black = "^22.10.0"
asyncio = "^3.4.3"
pyserial-asyncio = "^0.6"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 119

[tool.autopep8]
max_line_length = 119

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
